--- .env.dev
@@ -1,1 +1,0 @@
-ENV_TYPE= "dev"
