--- .env.prod
@@ -1,1 +1,0 @@
-ENV_TYPE= "prod"
