+++ lib/Store_App.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+// ignore: camel_case_types
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+      ),
+      home: const Scaffold(body: Center(child: Text('Hello, World!'))),
+    );
+  }
+}
