--- lib/screens/no_network_screen.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class NoNetworkScreen extends StatelessWidget {
-  const NoNetworkScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Container(
-          constraints: const BoxConstraints.expand(),
-          decoration: const BoxDecoration(
-            image: DecorationImage(
-              image: AssetImage('assets/images/core/no_network.jpg'),
-              fit: BoxFit.fill,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
