--- lib/style/images/app_images.dart
@@ -1,6 +1,0 @@
-class AppImages {
-  static const String noNeteowk = 'assets/images/core/no_network.jpg';
-  static const String empty_screen = 'assets/images/core/empty_screen.png';
-  static const String page_under_build = 'assets/images/core/page_under_build.jpg';
-  static const String splash = 'assets/images/core/splash.jpg';
-}
