+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,10 +5,8 @@
 import FlutterMacOS
 import Foundation
 
-import connectivity_plus
 import firebase_core
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  ConnectivityPlusPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlusPlugin"))
   FLTFirebaseCorePlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseCorePlugin"))
 }
