+++ pubspec.yaml
@@ -35,11 +35,6 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   firebase_core: ^4.1.1
-  very_good_analysis: ^10.0.0
-  connectivity_plus: ^7.0.0
-  flutter_dotenv: ^6.0.0
-  flutter_screenutil: ^5.9.3
-  flutter_launcher_icons: ^0.14.4
 
 dev_dependencies:
   flutter_test:
@@ -63,10 +58,10 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
-  assets:
-    - .env.dev
-    - .env.prod
-    - assets/images/core/
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
